Refactor profile model structure to improve package organization and naming consistency
